Added a basic persistent history
The shell will write each Readline::HISTORY entry to the Urchin history
file (by default ~/.urchin.history). Shells read this and populate the
history when they start.

This is a very basic implementation and can be improved upon in several
ways. Currently, the file buffer is flushed after every command. It may
be better to use a file in /dev/shm and periodically write that file to
disk. Also, shells may want to timestamp of identify commands in other
ways.