Shortcut STDIN methods and ability to pass extra Ruby params
Extra parameters can be passed to the Ruby process like so (given a
Shell.ruby_delimiter of "~@"):

  ~@-vp puts ~@

Spaces in the parameters are not allowed.

lib/ruby_require.rb is now required by the Ruby process. This currently
provides shortcuts to read STDIN into an Array, a String and an Integer.
:o is now an alias for :puts.

Tested-by: Mark Somerville <mark@scottishclimbs.com> [F14, MRI 1.8.7/1.9.2]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>