Allow recording of arbitrary data in the history
History is now a series of OpenStruct instances. This allows users to
record arbitrary data, by writing hook in ~/.urchin.rb. By default
Urchin will record the typed input line and the time it started. To also
record the working directory and Ruby version, you could write a hook
like this:

  History.before_appending do |history_line|
    history_line.working_directory = Dir.getwd
    history_line.ruby_version = Shell.new.eval("which ruby")
  end

This is the first step in addressing issue #14. Currently, there is no
builtin for accessing the stored data.

Unfortunately, storing this does come at a cost - the history file is
now binary data (the history entries are marshalled). If you find this a
bit disturbing, you can offset this somewhat by additionally storing
this data yourself in the hook.

Tested-by: Mark Somerville <ms@pccl.info> [F17, MRI 1.9.3]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>