Add a history builtin
21dc8f added the ability to record arbitrary information in the shell
history. This builtin allows access to that stored data in a way that
aims to be easy to manipulate with other tools. For example, to display
the working_directory and ruby_version and inputted line fields (that
were stored in the commit message for 21dc8f), separated by two @@s:

  history -o working_directory,ruby_version,input -s "@@"

This closes #14.

Tested-by: Mark Somerville <ms@pccl.info> [F17, MRI 1.9.3]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>