Don't blindly close FDs after fork()
Simply closing FDs > 2 after fork was causing problems on MRI
2.0.0-preview1, which uses a pipe internally to handle signals. It's not
clear why 1.9.3 wasn't exhibiting this behaviour.

Since the FDs that are reserved by the VM may change and because we use
1.9 or later (which support close_on_exec), Urchin now relies on this
rather than manually closing open descriptors. I realise that not all
kernels support this, but I feel that is an edge case that can be dealt
with if the need arises.

Tested-by: Mark Somerville <mark@scottishclimbs.com> [F17, MRI 1.9.3/2.0.0-dev]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>