Revert "Define Shell.subshell to run a command string"
This method of eval-ing Urchin code in-process was a blind alley. As
well as being mis-named, Shell.subshell was really a way to avoid
passing a Shell instance around.

In the end it turned out to cause more problems that it fixed, such as
the JobTable sometimes now being defined when the subshell received
SIGCHLD.

This reverts commit 92bd0d8a2b8b6aced7f81805ba994af03bf14fb6 as well as
changing any code that depends upon Shell.subshell to use an already
existing instance of Shell.

Tested-by: Mark Somerville <mark@scottishclimbs.com> [F17, MRI 1.9.3]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>